Created overloaded PacketType for temporary in/out detection
PacketType enum derefs into DataPacket, but stores context for putting
the packet into the correct buffer.

So a DataPacket is created with Sent if the packet was Sent from this
device, then when an ack is Received, the RTT of the Sent packet is
updated, but if we ignore the direction, the perceived direction of the
returned datapacket (the previously sent packet) is "Incoming", so
instead, we just temporarily mark the packet as Sent, and it is added to
the correct buffer.